Move code to respective folders for readability
